backend_classification/requirements.txt
# backend/requirements.txt

fastapi
numpy
opencv-python
pillow
python-multipart
uvicorn
uuid
